LPD-22028 Move PerformanceTestTimer under 'performance' package, use ItemCountPerformanceTestTimer as a private class for now
