+++ modules/apps/change-tracking/change-tracking-test/src/testIntegration/resources/META-INF/module-log4j.xml
@@ -2,6 +2,6 @@
 
 <Configuration strict="true">
 	<Loggers>
-		<Logger level="INFO" name="com.liferay.portal.kernel.test.util.PerformanceTestTimer" />
+		<Logger level="INFO" name="com.liferay.portal.kernel.test.performance.PerformanceTestTimer" />
 	</Loggers>
 </Configuration>
