+++ modules/apps/headless/headless-batch-engine/headless-batch-engine-test/src/testIntegration/resources/META-INF/module-log4j.xml
@@ -3,6 +3,6 @@
 <Configuration strict="true">
 	<Loggers>
 		<Logger level="INFO" name="com.liferay.headless.batch.engine.resource.v1_0.test" />
-		<Logger level="INFO" name="com.liferay.portal.kernel.test.util.PerformanceTestTimer" />
+		<Logger level="INFO" name="com.liferay.portal.kernel.test.performance.PerformanceTestTimer" />
 	</Loggers>
 </Configuration>
