+++ portal-test/bnd.bnd
@@ -7,6 +7,7 @@ Export-Package:\
 	com.liferay.portal.kernel.test.constants,\
 	com.liferay.portal.kernel.test.context,\
 	com.liferay.portal.kernel.test.junit,\
+	com.liferay.portal.kernel.test.performance,\
 	com.liferay.portal.kernel.test.portlet,\
 	com.liferay.portal.kernel.test.randomizerbumpers,\
 	com.liferay.portal.kernel.test.rule,\
