+++ portal-test/src/com/liferay/portal/kernel/test/performance/PerformanceTestTimer.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.kernel.test.util;
+package com.liferay.portal.kernel.test.performance;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
