--- portal-test/src/com/liferay/portal/kernel/test/util/ItemCountPerformanceTestTimer.java
@@ -1,95 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.kernel.test.util;
-
-import com.liferay.petra.string.StringBundler;
-
-import java.nio.file.Path;
-
-import org.junit.Assert;
-
-/**
- * @author Vendel Toreki
- */
-public class ItemCountPerformanceTestTimer extends PerformanceTestTimer {
-
-	public ItemCountPerformanceTestTimer(
-		int count, Class<?> clazz, String name, long maxTime) {
-
-		this(
-			count, getInvokerName(clazz, name), System.currentTimeMillis(),
-			maxTime, null);
-	}
-
-	public ItemCountPerformanceTestTimer(
-		int count, Class<?> clazz, String name, long maxTime,
-		Path logFilePath) {
-
-		this(
-			count, getInvokerName(clazz, name), System.currentTimeMillis(),
-			maxTime, logFilePath);
-	}
-
-	public ItemCountPerformanceTestTimer(int count, long maxTime) {
-		this(
-			count, getInvokerName(null, null), System.currentTimeMillis(),
-			maxTime, null);
-	}
-
-	public ItemCountPerformanceTestTimer(
-		int count, long maxTime, Path logFilePath) {
-
-		this(
-			count, getInvokerName(null, null), System.currentTimeMillis(),
-			maxTime, logFilePath);
-	}
-
-	public ItemCountPerformanceTestTimer(int count, String name, long maxTime) {
-		this(
-			count, getInvokerName(null, name), System.currentTimeMillis(),
-			maxTime, null);
-	}
-
-	public ItemCountPerformanceTestTimer(
-		int count, String name, long maxTime, Path logFilePath) {
-
-		this(
-			count, getInvokerName(null, name), System.currentTimeMillis(),
-			maxTime, logFilePath);
-	}
-
-	@Override
-	public void close() {
-		long delta = System.currentTimeMillis() - startTime;
-
-		double speed =
-			(delta > 0) ? (double)(_count * 1000) / (double)delta : Double.NaN;
-
-		log(
-			StringBundler.concat(
-				"Completed ", name, " in ", delta, " ms, speed: ",
-				String.format("%.2f", speed), " items/s"));
-
-		Assert.assertTrue(
-			StringBundler.concat(
-				"Completed in ", delta,
-				"ms, but the expected completion time should be less than ",
-				maxTime, "ms"),
-			delta < maxTime);
-	}
-
-	protected ItemCountPerformanceTestTimer(
-		int count, String name, long startTime, long maxTime,
-		Path logFilePath) {
-
-		super(name, startTime, maxTime, logFilePath);
-
-		_count = count;
-	}
-
-	private int _count;
-
-}
